// src/app/enums/panicEffects.ts
export enum PanicEffects {
    LaserFocus = 'Laser Focus = Advantage on all rolls for the next 1d10 hours',

    MajorAdrenalineRush = 'Major Adrenaline Rush = Advantage on all rolld for the next 3d10 minutes',

    MinorAdrenalineRush = 'Minor Adrenaline Rush = Advantage on all rolls for the next 1d10 minutes',

    Anxious = 'Anxious = Gain 1 Stress',

    NervousThitch = 'Nervous thitch = Gain 2 Stress. The nearest crew member also gains 1 Stress',

    Cowardice = 'Cowardice = Gain 1 Stress. For the next 1d10 hours you must make a Fear Save to engage in combat or else flee',

    Hallucinations = 'Hallucinations = For the newxt 2d10 hours(determined secretly) you have trouble distinguishing between reality and fantasy',

    CrippingFear = 'Cripping Fear = Gain a new permanent phobia. Whenever you encounter this phobia make a Fear Save at Disadvantage of gain 1d10 Stress',

    Owerhelmed = 'Owerhelmed = Gain 1d10 Stress',

    Rattled = 'Rattled = Let out a blood-curdling scream! Disadvantage on rolls for 2d10 minutes',

    Paranoid = 'Paranoid = For he next 1d10 days, whenever a character joins your group make a Fear Save or gain 1 Stress',

    DeathDrive = 'Death Drive = For the next Xd10 days(where  =Stress) whenever you encounter a stranger or known enemy you must make a Sanity Save or else immediately attack them',

    Catatonic = 'Catatonic = Become unresponsive and unmoving for Xd10 minutes(where  =Stress)',

    Broken = 'Broken = Fort the next Xd10 days(where  =Stress) make a Panic Check whenever a nearby crew member fails a Save',

    Psychotic = 'Psychotic = Immediately attack the closest crew member until you do at least 2d10 Damage. If there is no crew member nearby you attack the enviroment',

    CompoudingProblem = 'Compouding Problems = Roll twice on this table(page 8)',

    DescentIntoMadness = 'Descent into Madness = Gain 2 new phobias! Your Stress cannot be relieved below 5.',

    PsychologicalCollapse = 'Psychological Collapse = You become permanently, irreparably insane, Your character is now played by the GM',

    HearthAttack = 'Hearth Attack = Instant death',
}